reworked SDE items to use SDE2.

+++ model/sde/types-compiler/src/main/java/fr/guiguilechat/jcelechat/model/sde/hierarchy/AttributeDetails.java
@@ -1,12 +1,14 @@
 package fr.guiguilechat.jcelechat.model.sde.hierarchy;
 
+import java.math.BigDecimal;
+
 public class AttributeDetails extends CommonDetails {
 
 	public boolean hasFloat = false;
 
 	public boolean highIsGood;
 
-	public double defaultValue;
+	public BigDecimal defaultValue;
 
 	public boolean stackable;
